Renamed folder for Agora resources
